+++ public/app/plugins/datasource/influxdb/specs/influx_query_specs.ts
@@ -15,6 +15,18 @@ describe('InfluxQuery', function() {
     });
   });
 
+  describe('render series with policy only', function() {
+    it('should generate correct query', function() {
+      var query = new InfluxQuery({
+        measurement: 'cpu',
+        policy: '5m_avg'
+      });
+
+      var queryText = query.render();
+      expect(queryText).to.be('SELECT mean("value") FROM "5m_avg"."cpu" WHERE $timeFilter GROUP BY time($interval) fill(null)');
+    });
+  });
+
   describe('render series with math and alias', function() {
     it('should generate correct query', function() {
       var query = new InfluxQuery({